DB: support for prepared statements.
Whenever some code would repeatedly call the same, complex, SQL query,
it makes some sense to prepare that as a statement.
However, prepared statements only last for the duration of one postgres
client connection, so the Cursor code should keep track of the ones that
are sent to the server.