ir.translations: rewrite indexes, hash for src.
There is a rare chance that a translation's source text is more than 1.3k
characters, which will then not fit a b-tree page. Use hash indexing for
the source.

Moreover, Postgres suggests that a combination of two indexes can work
quite well, instead of having a multicolumn one for every combination
of columns.