ORM: reduce multiple calls for ir.ui.view queries.
In fields_view_get, we need to discover all the views that could apply
to the object. In postgres 8.4, this can be done using a single query.

This patch did not introduce any great saving, since the bottleneck
is the xml processing. However, it could theoretically reduce the db
calls, especially if the db were remote.