XML-RPC: initial support for http authentication
This is the alpha code for allowing http-based authentication, under
a '/xmlrpc2/' path. With http-based authentication, we could avoid
sending username/password for each rpc call in the same session. However,
on a new session, the protocol will respond with 401 and require an
authentication for the first call.
Note that xmlrpc2 won't make sense for non-persistent connections.
This change should NOT affect the old, xmlrpc, protocol.

Implementation notice: Since we want to issue the 401 Authentication Required
response early enough, we need to know what domain (= realm) to ask
authentication for. So, we must distinguish between "pub" calls, that
require no authentication, "root" ones that need the admin password and
"db" ones that need a valid res.user for some specific db. Thus, we separate
the paths to the xmlrpc2 to '/xmlrpc2/pub/', '/xmlrpc2/root/' and
'/xmlrpc2/db', respectively. The authorization level depends *only* on
those paths, rather than the called method (which is in the POST body).
Then, the ExportService.new_dispatch() *must* check that the correct
authorization level (in auth.provider.domain) is supplied for some function.