sql_db: Fixes after merge, re-use connections.
The old logic of reusable connections had been broken wrt. transparent
reconnection. As soon as I fixed it, I discovered that bad connections
would not be discarded (thus persistent error) and prepared statements
wouldn't work (because the cursor is re-created, while the conn holds
the prepared statement). Fixed all that.