Res.company: fix searching for the default company logo.
The previous logic was $root_path/../pixmaps, like in a development setup.
Reading files outside our cwd is a bad idea, and if the pixmap couldn't
be found (like in a setup through packages), a new database would fail.
So, better to have a:

[paths]
pixmaps = /usr/path/to/pixmaps

explicit setting in the config file.