ORM: hack to enable browse_search API
A major feature, often requested, is to replace lines like:
    ids = obj.search(cr, uid, <expr>)
    for sth in obj.browse(cr, uid, ids): ...
with:
    for sth in obj.browse(cr, uid, <expr>): ...

Now, we enable exactly that, with some really trivial implementation.
Modules could adapt to the new API, while the orm core can be improved
to reduce the implementation into one query (some day).

Famous last words are that this patch is too trivial to go wrong.