http server: keep track of handlers, try to stop them.
The http server classes are quite complicated, because they form a
real mesh^Hs of objects to listen, thread and serve requests. Try
to keep a track of the multihandlers, in a list, so that upon exit
we can locate them, close their connection and hope that they will
finish their threads.