websrv_lib, netsvc, httpd: armor the code for clean server exit
There is many border cases we need to catch, namely a solid collection
of the handlers (http) that are in use and correct removal of those.

The goal is to have a clean (as possible) server exit, and that *all*
running servers (+threads) get a proper notification upon exit, so that
they can perform any cleanup they may need.