RPC: support for content-rich fault responses
https://bugs.launchpad.net/openerp/+bug/257581

Following that thread and discussion with devs, realized that the server's
XML-RPC v1 is abusing the faultCode,faultString elements to send the text
and traceback of an exception, respectively. It also uses some object
juggling, until the exception is really sent to the xml-rpc underlying lib.

The current v1 behaviour *must* remain, or else the native clients will
break wrt. displaying the exception. However, it does not follow the XML-RPC
standard, nor does it have a well-defined protocol.

Solution is to change that *only* for XML-RPC2. We reshape the exception
handling of the dispatchers, so that 3-4 fields of info can be passed down
from the dispatchers to the RPC protocol, in a well-defined API. Then, at
XML-RPC2 (and any future protocols), have all the important information
formatted into the faultString element.

Choose a simple format, resembling e-mail header fields. The faultString
will contain lines like "X-Field: value", which will then be parsed by the
client. This way, future extensions to the exception data are possible.

Under the hood, code for XML-RPC1, NET-RPC is improved. But, at the data
protocol, this version should produce the exact same responses that previous
servers would send at exceptions.