ir.model.access: cache check_groups(), consolidate it in ORM.
It was noted that frequent queries would look for the same groups
resolution (like check if uid belongs to base.group_extended). This
shall better be cached in python, to save the sql roundtrip overhead.

Clearing that cache is not trivial, because it depends on
res.users.group_ids and ir.model.data. So, workaround with a timeout.