Modules init: Load xml data optionally on dependencies
When a module loads its data, all referred models or data should exist,
or else loading will break. This would mean that the module must depend
on any other modules it needs to use the data of. However, sometimes we
just want some module to define elements of data for modules it need not
strictly depend on. Then, we could tell the xml loader to only load a
block if some modules are there:
	<data with-modules="foo,bar"> ...