Http: Support for gzip-encoding for xml-rpc data.
Obvious goal: to reduce the traffic and hence latency of XML-RPC calls.

Implementation: this feature should be truly transparent to the server
and clients, and backwards compatible for both.
We use the standard "Accept-Encoding" and "Content-Encoding" http headers
to mark gzipped data. We support gzip *both* at the POST request, and the
http body responses (most http implementations only care about the
downstream direction, the response, but we want to save bandwidth on both).

The server shall /always/ accept gzipped requests. It may ONLY send back
compressed data if the 'Accept-Encoding' header of the client allows it to.

The client shall ONLY send gzipped requests if the server indicates support
for them, through the 'xmlrpc-gzip' server option.