websrv_lib: try to avoid some exceptions on server exit
If client disconnects, we get a broken pipe or timeout, in which case we
should gently break the request. On server exit, the sockets close, so
the attempt to finish requests (sending data on them) will raise an exc,
that we would need not process any further (ie. try to send 500, too).