netsvc: Optionally log the dbname
We might want to log the database name when debugging multi-db issues,
but normally we *must* avoid making our server any slower wrt. repetitive
calls (like log() ). So, this code was written with the strict requirement
that when the feature is turned off, no overhead should incur to the
server. If turned on, a moderately elaborate way of determining the db
will be used, iff the log message is about to be printed at all.

Also, the dbname is presented as "loggername[@dbname]" so that it is
easily parseable.

No lemmings were harmed in making of this patch.