ORM: enhance the "fields_only" feature with 'auto' mode
Some months ago, the "fields_only" feature had been added to the orm
browse objects, so that they wouldn't fetch the unused columns from
tables. Fetching too little columns, however, meant that we would have
to repeat the query for the rest, so could be marginally be sub-optimal.
Now, have an "auto" mode that will try to count the columns popularity
and prefetch the ones that are mostly used.