customfonts: search path for fonts, override the broken of reportlab
Several Linux distros still ship with a broken reportlab config, which
looks at "c:\winnt\fonts" for fonts!
Since they have not fixed that for months, we have to do their job and
have a custom search path of:
   - paths from the config file (ttfonts.search_path key)
   - sensible defaults for distros, considering os.name and os.uname
   - default reportlab path (opt-out with ttfonts.use_default_path=False)

The result is that we will have more chance of locating TTF fonts and
use them (as now required) in the PDF reports.