ORM: improve columns change type/size during upgrade.
When a selection field is updated to contain a longer keyword, the size
of the corresponding VARCHAR field has to adapt. That's also true for
all char and varchar fields.

At 2009, Joël Grand-Guillaume had limited that algorithm to only
increase the size of VARCHAR fields. This patch will re-introduce size
decrease, when all data can fit the new smaller size.

Side-effect 1: obviously, sizes of fields will now be different
Side-effect 2: adaptation of data will be faster.
Side-effect 3: views are NOT dropped any more (no "CASCADE" of a column
DROP). This may be good, because views will persist[1], but also /bad/
because some view referencing those columns may cause an sql error.

[1] Say we have base-A, mod-B and mod-C inheriting in in that order. If
mod-B introduces a view on base-A's table and mod-C changes a column of
that view, old code would silently drop the view (and only restore it
after 2 upgrades). This code will raise an exception.