orm: csv import data mangled (with solution) [Bug 698354]
Many2one fields with name are getting related to the wrong objects with hilarious/disastrous results.
e.g. my res.users.csv specified menu_id of 'Menu' but got 'Board Create Menu' resulting in complete lockout :-(

The problem is name_search() is building this query:

   SELECT "ir_actions".id FROM "ir_actions" WHERE (ir_actions.name ilike
E'%Menu%') ORDER BY name limit 100

The use of 'ilike' means that if there is more than one action with 'Menu' in the name then it is pot-luck which one you get and more than likely chaos will ensue.
The match must be exact, so the operator must be specified as '=', thus:

=== modified file 'bin/osv/orm.py'