ORM: make sure we don't duplicate existing translations [Bug 705364]
When copying a record containing o2m, m2m fields, we (falsely[1]) traverse
down the structure and try to copy everything, including the translations.
Then, when some field is a function one, the algorithm would still
duplicate translations for records that haven't been copied.

Reported by: Sébastien BEAU <sebastien.beau@akretion.com>

This is an alternative solution, which checks if translations for the
destination records exist, before creating them. It shall work irrespective
of the field type.

[1] copying the data of the fields is not always the proper way to
duplicate a record. See pg84-next branch. This patch intends to be a
forward-compatible solution to that.