yaml import: allow to specify uid for the interpreter
This will pass on any uid to the interpreter, so that things are run as
that user.

SECURITY NOTE: this is *NOT* safer than admin, by any means!
Even if uid is passed down, the yaml code still has the ability to elevate
privileges and execute stuff. Do *NOT* trust users to upload YAML. This
feature is here only at cases where a controlled experiment needs to run.