ORM: copy() records virtually
If we have model B inherits from A and try to copy a record of A (which
is in fact an instance of B), do copy B for the resulting record.

This will only work on classes with virtual table enabled, so old-style
ones won't be affected. Ones that define virtual behavior, however will
yield a child-copy instead of just the baseclass.