http_server: refactor IPv6 support for easier configuration
Previously, we would have the "httpd", the "httpsd" and then the "http6d"
etc, each of those able to listen on one interface/port.
With the introduction of IPv6, we may have the need to specify a set
of interfaces to listen to (as opposed to listening to all), and these
would also conflict with IPv4 ones on the same port.

So, have a "multi-interface" ability in the "interface" setting of two
httpds, accepting both IPv4 and IPv6 syntax.
Example:
   [httpd]
   interface = [::72ff:fee2], 192.168.1.2, 10.0.4.1:8169

This shall be compatible with the previous (single i/f) syntax, but
also allow configurations where we really want to restrict access to
trusted interfaces (networks).