orm: Allow call of model.browse(cr, uid, ids=<browse_record>, ..)
This may unlock an important optimization for the API: When we call
some_model.browse(), we allocate a new cache where fetched data are
stored. Sometimes we are already in the context of some browse()ing,
so the cache (with our data) could be reused.
In fact, we end up doing (in a few calls, simplified here):
    for bro in self.browse(cr, uid, ids):
        res_a[bro.id] = bro.alpha
        res_b[bro.id] = call_foo(cr, uid, id=bro.id): #(expanded here)
            return self.browse(cr, uid, id).beta
.. which means that 2 separate read() calls would be made to retrieve
model['alpha'] and model['beta'].
Now, the inner function call_foo() will be allowed to reuse the browse
record and its cache, transparently with pretty-much the above code.
(in that case, let "id=bro", instead of "id=bro.id")

This feature needs to be explicitly enabled wherever it makes sense.