[FIX] orm.browse_record: fields prefetching should ignore m2o pointing to models not loaded yet
This situation can occur due to custom fields (x_...) that are creating shortcuts between
objects without respecting the module dependencies. For example a custom m2o fields 
connecting a partner and a product would fail to be loaded at server startup when only
base is loaded, because product.product is not yet available.
This can usually be simply ignored, as the prefetching is just an optimization.
The custom fields should never be accessed during startup (e.g not function field
should use them), so that should never cause any issue.