ORM: let orm_memory models have custom fields [Bug: 731487]
This is a _feature_ patch.
Reported by: Sébastien BEAU @ akretion.com

Until now, creating custom fields in orm_memory objects (like x_foo)
was not really supported. The _columns of the model wouldn't be updated.
Introduce this functionality, and also fix renaming of the _columns for
them.

Note, we still /don't/ support reading the new column of a record that
had been created before the column was added from the GUI. Since orm_memory
records are volatile, this functionality may be unnecessary.