ir.model: when validating ir.model.fields, exceptions must rollback.
This is a general practice that may be missing from other parts of the
code, too. When doing post-write() validation, raising an exception will
not guarrantee that the data of the write() will not be stored in the db.
The cursor is still open and valid, and the calling code can still commit
it.

So far, we have been protected by the fact that clients do manipulations
in several transactions. However, in the case of
    extra-addons/test_orm/test/orm_fields.yml
that didn't happen and the data was actually committed.

Beware of the opposite scenario, too, where an exception would invalidate
the cursor (or explicitly roll it back) where a partial failure could be
tolerated (is it ever?).