tools/server-check: inverse the trigger-file logic
Previously, we had used the empty file to indicate that we need the check.
However, in Mageia, /var/run is cleaned upon system boot, so this logic
couldn't work.
Use the opposite, that is, have a "checked" file to indicate that we
don't need to check again. Hope it works now.