ir.translation: select restriction must be an SQL constraint
select fields validate themselves in ORM, through _check_selection_field_value(),
which /retrieves/ the .selection list and checks if value to be written
is among possible ones. This, however, is very expensive for models that
have variable selection lists (eg. ones that require an SQL query).

In fact, for ir.translation, the query would be repeat for every single
write() to ir.translation. The 'lang' field is actually a reference for
res.lang.code (not the .id, sadly), so let SQL implement the constraint
for us.