orm: factor orm._update_store() to group UPDATEs together
This optimises the special (but frequent) case that several of the
computed values are identical. In that case, we'd better detect
that and call "UPDATE SET val=xx WHERE ids = ANY(...)" . Hopefully,
postgres will save some time with this batch update.

Note that _update_store() is not called during every-day orm traffic.