orm: respect Postgres schemas
A feature, blame me if it makes your cat sad.

It is a common practice[1], when we want to hide older tables in postgres,
to move them to a different schema, which is not in the 'search_path'.
Then, these tables disappear from regular queries.

However, the ORM, when installing/upgrading a module, looks up the tables
by name only, irrespective of schema. So, it would find them, assume they
are available, and not create them again in the 'public' schema.

Now, put an extra sub-query (costs 2sec in the all-modules test) that
limits pg_class searches to the "visible" schemas.


[1] when migrating data ;)