[IMP] fields.function: removed method=True param, added docstring
The 'method' param was quite useless, as 100% of functions
fields were using method=True. In addition, there is no need
to distinguish methods and functions, as methods are unbound
and passed as function objects in the declaration of a function
field, so they only need a proper signature.
Finally, docstring was added for fields.function class,
based on current doc from developer book (in preparation of
future API doc).
(A fix in addons follows, getting rid of all the useless
method=True params there too).