ir.model.data: require "noupdate", introduce 'source' column
Triggered by the new ORM engine, which put a 'DEFAULT false' attribute
in i.m.d .

Until now, ir.model.data.noupdate column could silently accept a NULL
value, having /almost/ same meaning as False. The big difference is that
a NULL would escape the check at _process_end() and therefore not
have these records *deleted*.
Side-effect of storing the default False value in SQL was that models
and model-fields would now be checked, because of that value. See, these
ids had been explicitly (for speed) inserted into the db bypassing the
ORM (which would set the noupdate value).

Instead of keeping the NULLs (which are incompletely modelled, anyway),
an old request came to solve this issue: introduce the "source" column,
also provisioning for non-XML originated records (see
base_external_referentials).

Note that a many2one column referring to a "source" table is IMHO a
bad idea. ir.model.data is a table we want to keep as fast as possible,
so an extra foreign key might reduce performance.