base: bump version, add migration scripts
Since ir_model_data and ir_model_fields are used early in the upgrade
process, extra columns must be added with a migration script.

Note that these include the "field_data" column, not yet appeared in
the ORM. That's a provision, so that I don't need to bump the version
twice.