expression: fix parsing for indirect (function) fields
A regression in the recent "parse_into_query" algorithm was that left
expansion would only happen once, at the first stage of parsing. For
regular fields, this is desired, as we mandate the `expr_eval()` to
return only columns that exist in SQL.
However, function fields could return a "rewritten" expression in terms
of the other model's fields. These may need a second parsing, in order
to convert them to SQL fields. Example is a related field which returns
an expression on a one2many column.

With this patch, the first stage is separated, again[1], into the
__cleanup(), at the expense of performance[2]. The intermediate
'dirty_expr' class indicates that some 'expr_eval()' result[3] needs
to be cleaned again.

[1] the old "parse()" algorithm was using 2 stages, too.
[2] I could have just copy-pasted this loop into __cleanup() and have
the same code twice. Faster but more error-prone in the long term.
[3] since the refactoring, any solution like "if type == 'function'"
is unacceptable. Code in 'expression.py' must be agnostic.