fields.string: Only translate if lang in context
Warning: changes ORM results!

That block of code meant: when trying to match a string of a
translatable field, look at the current language (from context)
_OR_ at the en_US translation.
Well, let's get rid of that. If there is no context language, we cannot
assume that en_US needs to be searched. Our language is "C"[1], we may
only search that one.
Practically, this introduces a significant speedup of several char/text
fields.

[1] every time we mention that, we pay tribute to Dennis M. Ritchie. RIP