netsvc: defer and fix pretty_repr() of Agent function
There was this non-trivial chunk of code to print the called function in
a nice format. Two defects: it would not accept non-class functions, and
it would compute even when debug had been turned off.

Instead, move to a class where evaluation will be deferred, support
functions with global scope.