tools: Introduce 'server types' for safe context parameters
The idea is a bit old, the problem too: we need to pass some explicit
flags down the ORM calls (through context), but don't want them to be
available to RPC callers. They must only work locally.

The solution is simple: use special types, that cannot be marshalled
through RPC. Emulate the basic types of python.

Also, have convenience functions that will check the values, in particular,
inside the context dictionary.

Note: marshalling is not fully prohibited now; the values will be mapped
to their standard type equivalent. To be fixed/discussed.