res.partner: convenience function to browse the best address
We need it all the time: we need to find the correct address of a
res.partner to use. We had 'address_get', but this would return /all/
matching addresses, just by id.

Instead, have another method that returns a _browse object_ of the best
matching addres (non-empty, matching type and validity).