http_server: enable TCP keepalive on HTTP connections
This should clear long-hanging connections, that may be the cause of
resource excaustion.

Go to an Internet Cafe, connect to our OpenERP instance (through https,
obviously). Then, remove the laptop without closing the client. That
connection will take days (TCP-wise) to be freed. Keepalive should tackle
that, but more tests need to prove that in practice.