workflow: major refactoring, new OO engine
The relic engine had survived since the pre-5.0 days of Tiny. It was about
time I revamped it. Unlike Tryton, support for workflows will stay in f3.

But, still, this code *does* implement the old engine, delivers exactly
the same functionality[1]. It is just reformed in a set of OO classes,
lives persistently in the server cache (instead of loading the wkf each
time a record is about to change).

[1] with the exception that previously gray-areas (like cross-wkf
    transitions) are now impossible.