orm, fields: "merge" functionality. API
In 2009, two modules had been written: account_invoice_merge and
crm_case_merge, which partially tackled the issue of duplicate records.
AFAIR there was also a discussion with Dr. Gassauer, back then.

The latter algorithm (for crm.case) was generic enough to inspire a
flexible solution.

The problem is, the merge "rules" depend on the model fields, which in
turn may depend on multiple modules inheriting and altering fields. So,
rules have to be managed at the framework level (not a module), just
like the fields are.

So, put them where they belong, try to have sane but strict defaults.

This is alpha quality. Important TODOs pending.