sql_db: allow 'temporary' connections, for special operations
Triggered by the last commit by N. Bessi, which would be sub-optimal,
and the doc_ftp issue in buildbot (which SA. refuses to fix), an
addition to the API is introduced.

Sometimes we want to use the sql_db API to temporarily connect to a
database (such as 'template1', or a real one), and then release the
connection. Have a flag to indicate so.