osv: allow to delete non-existent columns
At commit 15a71b4afc5610d70 we let None remove columns at inheriting
models.

This feature is useful to control the order of inheritance. Say, model
B inherits from A and adds column 'foo'. Then C inherits (in a different
module) from A, but the order is not guaranteed, so 'foo' may or may not
exist.

A -> B.foo => A.foo \
  -> C     => C ?     => C.foo ?

in this case, we can say C['foo'] = None, meaning that 'foo' will not
be inherited.

However, in that case, C may load earlier than B, so 'foo' will not be
available.
Well, tolerate that!
Use 'pop(, None)' rather than 'del', which raises no KeyError.