orm: Crude fixes for the merge_records() functionality
Ignore many2one field, if that is on a orm_memory model[1] or a view[2]
Add stub handler for relational fields

Add an entry in ir.model.data for every record merged. This will help
in the future, if that missing id ever needs to be traced again

[1] should we dive into the data and update them, better?
[2] that's not entirely correct, as non-_auto models may be customised
tables, theoretically. This shall be revised once we have a dedicated
orm_view class.