base/security: ir.model.data must be writable for merge records
Merge records leave one entry in ir.model.data, about the record that has
been removed. This, however, needs write permission on i.m.d, so create
a special group for that and grant access.

This is a bit dangerous, in terms of performance. Read access to i.m.d
is requested all the time, and the rule shall not hinder its speed.