orm,fields: per-field ACLs, as in older API
Ever wondered what these "groups" stood for in ir.model.fields?

It had been a feature in the pre-5.0 days, when group restrictions could
be specified for individual model fields. Broken since then.

The ACLs is a good idea, often requested by many OpenERP integrators.
This commit brings it back to life, using an otherwise lightweight
implementation. It also fixes the ir.model.fields.write() so that these
ACLs can be configured on-the-fly.