workflow: introduce pre_write() call before orm.write() is computed
Feature, API change.

In the old engine, workflow transitions would only be triggered *after*
all values had been written into the ORM record. So, no chance of a
trigger like "when model.state goes from 'draft' to 'open' " being
implemented.

With this API change, the workflow engine is called with the old values
still at the record. Then, it should test for transitions and fill a
dictionary of record ids to signal.