http_server: log the not-found 404 paths
When we reply to the http request, we only need say "Not found" rather
than copy the request path.
But, at server side, it's useful to see what was that path.

Note: the severity of the message may need to be revised. Since the
request may come from any unauthenticated source, it is always possible
to flood our logs with bogus paths. If so, the 'httpd' logger level
can temporarily be tweaked, or even this line changed to some other
level.