web_services: specify Postgres template to create dbs from
Several months ago, Nicolas Bessi of Camp2Camp had hacked the DB
creation code to automatically pick "template_postgis" rather than
"template0", if the first were available.

Now, we need a more generic way of selecting the template database
to create from. Reason is, some features (like 9.1 extensions, langs,
PostGIS etc.) must be prepared by the postgres superuser *before*
some OpenERP-F3 modules are installed.

So, for example, we could create a "template_openerp" database, fill
it with extensions and features, and let that be used, by setting in
the config file:

  [databases]
  template = template_openerp

Note that we could also specify multiple entries, like:

  [databases]
  template = template_openerp template1 template0

in the order we want them to be preferred.