fields.id_field: helper for data synchronization
The ir.model.data in F3 will now hold external references in a
defined API.

Say, among the fields of ir.model.data we will use:
  source = 'sync'
  module = <the module performing the sync>
  name = <ext_server>+':'+<ext_id>
Where the "server:id" syntax may further be adapted to your needs.

For this API, a helper is written under the 'id' field so that we
can efficiently select the records of some model that are in or
out-of-sync with the remote database:
   res.partner.search([('id.sync.magento', '=', True)]) # for those
                                                        # in-sync
or sth. like:
   res.partner.search([('id.sync.magento', 'not like', 'dbhost1:')])
   # for those who are out of sync wrt. "dbhost1" external db

Behind the scenes, the ORM will locate records which are (or aren't)
referenced in ir.model.data with adequate timestamps.