ir.model.fields: message when ORM models are inconsistent
On my oldest db, tried to do a res.partner.merge_records(), which
would need to "touch" all models referencing res.partner. That's
a hell of a consistency test, in fact!

Rather than a cryptic KeyError, raise an AssertionError explaining
that a model is missing some column.
This happens because ir.model.fields in SQL are not aligned with
the fields actually in the pythonic definitions of the models.