addons,ir.model: more logs when removing a module
It has happened that an incomplete db installation wouldn't have
all the ir.model.data entries for the base modules of the one I
was trying to uninstall. In such a case, the ir.model.field of the
base class (which should have been preserved) was attempted to be
removed.

With this patch, we give enough logs to identify such a problem,
and/or similar situations.