sql_db: optimize by avoiding psycopg1 cursor
The algorithm for dictfetch*() of psycopg1 cursors is notably
sub-optimal. We can do better than that,  evaluating the names
of columns and then dict(zip(names, row)) in a loop. Note that
this turned also to be faster than a C experiment, because the
dict(zip()) operates on pre-computed lists of names+col_values.