tools/service_meta: Forward-inheritance metaclass
The idea of forward-inheritance is not new, it is what the 'osv'
has been doing all along. Here, we formalize this behavior in a
metaclass. So, we can apply to all cases we want to dynamically
modify some class with a plugin.

So far, the plan is NOT to apply the new style to the ORM API,
but only to classes/features introduced in F3 (ie. after 6.0).
This way we respect the known API of `osv`, that happens to be
working well enough for us.