addons/base: servicemeta of "ExecContext"
It is common in several plugins[1] that we want some explicit context,
aka. locals/globals, for the "eval()" or the "safe_eval()". This can
be formalized in the "ExecContext" class, which builds the context
and can be extended through addon modules.

The "ExecContext" belongs to the 'base' module, because that is the
common denominator dependency of the modules we wish to use it in.


[1] think of "base_messaging" and "base_template_jinja2" for a start.