[IMP] res.users: enable cache on context_get, used by the _() translation function
When _() fails to find a context lang it looks
up the user language by calling context_get().
During module installation this may happen
thousands of times, and drastically slow down
the installation. Caching context_get calls
is an effective way to work around this,
until we find a better solution.