http_server: use weakrefs to keep track of threads and handlers
There is a suspected, slow leak of memory for MultiHTTPHandlers,
probably when some connections don't finish() normally. We had
been keeping a list of all active threads and one of handlers, so
that a server shutdown would close them. Keeping them in weakrefs
should be safer, since they will immediately be free()d as soon
as a request/connection is closed. Then, we only need to clean up
the weakrefs list occasionally.