res.user: remove the "welcome email" functionality
This had been a complicated feature, implemented too early in the
base module. Perhaps the only reason smtp binding would be needed
before proper setup of a server. Not even translatable!

Instead, some module should re-implement it with a hook to res.user
and using the base_messaging framework. Definitely not here, in the
server repository.