[FIX] base: invalid address in demo data causes failure in Google Maps lookup
Apprently on 2013-03-08 Google Maps stopped
returning a location in Belgium for this invalid
street name.

We consider Google Maps as a commodity so it
is ok to have the crm_partner_assign tests
fail in case our lookup mechanism does
not work anymore.

Note: this is a cherry pick of the 7.0
fix as a quick fix because the complete
forward-port of 7.0->trunk will require
some more time for conflict resolution