sql_db: optionally activate the binary SQL cursors
This depends on an experimental feature of psycopg2, it asks the
library for a "binary mode" cursor.

We're finally here! we can really test passing the data in binary
mode, through PQExecParams() call. Even if speed is hard to measure,
this patch will reveal all places where typecasting was naive in
the addons code.

Pending A LOT OF testing, improvements in psycopg2 etc.