sql_db: inherit ConnPool from Service, let old connections expire
Instead of a dummy object, let ConnectionPool be a Thread of Service,
which will allow it to run a periodic loop and close connections that
have been unused for some time.

This should theoretically free up resources of our server machine,
as Postgres does allocate a process for every connection opened.