websrv_lib: HTTPModified mixin to handle If-Modified-Since
In several http services, we will gain significantly if we manage
to exploit the If-Modified-Since header. One of them is the simple
HTTPServer, found in this base library.

So, the mixin cannot be later than here. We export two functions,
that test for the header and then compare to our "target" tstamp.