[FIX] res.partner: make use_parent_address field visible if set, even if parent is missing
This should not be needed but in the initial 7.0 branch
the default for use_parent_address was True so it was
set even on partners that had no parent company.
Now that the fields are read-only when that option
is enabled, it makes the address impossible to
edit. At least by making it visible we give the
user the opportunity of fixing it manually.

Can also be fixed at once with a single SQL
query:
UPDATE res_partner set use_parent_address = false
       WHERE parent_id IS NULL;