tools.convert: major refactor, split into modular "loaders"
The monolithic tools.convert is no more!

Replaced by a modular "service-meta" set of loaders, which can be
extended. At the same time, done a big cleanup of the old code,
remove much of the duplicate "eval xml" and also unreachable code
that stems back before the 5.0 era.

The new code bears more symmetry, and is also (a bit) more documented,
in particular wrt. the available elements and context of eval() calls.