orm: fix _check_recursive() for browse and models with "path" field
First, that code would break if passed browse_list() instead of plain
ids in a list. This is expected to happen for _function_field_browse
models and is legal.

Second, if the model had a "path" or "cycle" named field, the SQL
code would become ambiguous. Use more explicit names.