fields: a _defaults fn() may not work during module upgrade
There is some _defaults functions that depend on dynamic conditions
and user context (eg. get user's current foobar), which means that
they could fail on an automatic upgrade cycle. During such a cycle
we would have run:
  UPDATE foo.model SET somefield = _defaults.somefield(cr, uid...)
        WHERE somefield IS NULL

Calling that function would fail, however. So, mark them with an
attribute to avoid such a scenario.

The API now is:

  def _some_fn(self, cr, uid, context):
     ...

  _defaults = {
     'somefield': _some_fn,
  }

  _some_fn.no_auto = True