[IMP] tools: improved html_email_clean.
It can now take into account sections, and protect them. This allows to display a read more link
after a whole section. This will be used in the website, to display the first section of blogs for
example.

Also added some options on the read more link, allowing to tune its tag, content, as well as the link.

Also improved html_sanitize, taking improvements from the trunk branch of server, waiting for the whole
trunk branch to be merged.