sql_db: tolerate unicode in SQL queries errors
When SQL queries fail, postgres errors will echo us back some
of the data parameters[1] of the query. If these contain utf8
characters, then logging of the error fails. This also breaks
recognition of constraint errors by the OSV layer.

This commit /breaks/ the formatting of SQL errors in the logs.