ir.actions: don't serve search_view for a form view
The ir.ui.view code has a function field on "search_view". It
does an extra query to fill the ID and may even perform an
expensive operation to compute the view if no explicit search
view is defined for the model.

But, when the ir.ui.view type == 'form', no search view is
needed at all. Avoid the expensive operation!