ir.model.data: always check for source=='xml', res_id != 0
The convention that all ir.model.data come from module XML files,
or that all `res_id` values are valid is soon to be broken. Thus,
we need to always define what source we want to search for[1] and
also that we want a valid (non-zero) res_id.

Checking that `res_id != 0` when `source = xml` is also redundant,
it will not ever happen. But we'd better keep explicit on that.

[1] for some time now, the index has had source as the first part,
so we had to specify it.