http_server: block XML-RPC[2] calls with Cookie auth
Since "token" (aka Cookie) authentication was introduced, it was
possible to use it against XML-RPC. But no proper CSRF protection
has been implemented yet. So, block it until the protocol is
amended and pepper is added to it, too.

Note that this code may only change a little, just to feed the
actual pepper to 'checkPepper()'