netsvc,osv: special handling of Access Errors, without traceback
The OSV layer has long had a catch for `except_orm('AccessError', ...)`
but it would generate the usual RPC error, with traceback.

Now, we leave the "AccessError" string untranslated (note, no space),
let OSV handle it differently and yield an RPC error w/o traceback.

You are /advised/ (but not requested) not to translate "AccessError"