translation: when exporting translations, don't filter ir.ui.menus
ir.ui.menu `search()` is hacked to only return the "visible" menus[1],
but that means that some of them may be missing from the translation
catalogues.
Put a flag in the context[2] and let all of them be exported

[1] this should have been done the other way round, ie. the Gtk client
only asking for the "visible" menus, through the context. However, the
current implementation (of the client(s)) works this way and we must
thus adapt the server.

[2] actually, two flags. One for the current ir.ui.menu `search()` and
one more for future models that may work the same way.